test: cover rating stats visibility
